+++ src/CryptoTracker/CryptoTracker/Import/Objects/BinanceWithdrawal.cs
@@ -19,12 +19,12 @@ namespace CryptoTracker.Import.Objects
         /// <summary>
         /// Preis die ausbezahlt wurden. Entspricht Preis nach Fee Abzug.
         /// </summary>
-        public double Amount {  get; set; }
+        public decimal Amount {  get; set; }
 
         /// <summary>
         /// TransactionFee in der Coin-Symbol Währung
         /// </summary>
-        public double TransactionFee { get; set; }
+        public decimal TransactionFee { get; set; }
         
         /// <summary>
         /// Zieladresse
