+++ src/CryptoTracker/CryptoTracker/Import/Objects/MetamaskTransaction.cs
@@ -33,12 +33,12 @@ namespace CryptoTracker.Import.Objects
         /// <summary>
         /// Anzahl der Coins vor Gebührenabzug
         /// </summary>
-        public double Amount { get; set; }
+        public decimal Amount { get; set; }
 
         /// <summary>
         /// Preis die ausbezahlt wurden. Entspricht Preis nach Fee Abzug. 
         /// </summary>
-        public double TransactionFee { get; set; }
+        public decimal TransactionFee { get; set; }
 
         public string Kommentar { get; set; }
     }
